% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{abspath_read_and_unnest}
\alias{abspath_read_and_unnest}
\title{abspath_read_and_unnest}
\usage{
abspath_read_and_unnest(tibble_with_abspath_col, sheet_string = "label")
}
\arguments{
\item{tibble_with_abspath_col}{a tibble}

\item{sheet_string}{a string}
}
\value{
a tibble
}
\description{
abspath_read_and_unnest
}
