% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{abspaths_vector_to_col_in_tibble}
\alias{abspaths_vector_to_col_in_tibble}
\title{abspaths_vector_to_col_in_tibble}
\usage{
abspaths_vector_to_col_in_tibble(paths_vector, col_name = "abs_path")
}
\arguments{
\item{paths_vector}{A chr vector of paths}

\item{col_name}{Column name for paths_vector in tibble}
}
\value{
a tibble
}
\description{
abspaths_vector_to_col_in_tibble
}
