% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{add_pathogen_spore_conc_column_and_relocate}
\alias{add_pathogen_spore_conc_column_and_relocate}
\title{add_pathogen_spore_conc_column_and_relocate}
\usage{
add_pathogen_spore_conc_column_and_relocate(
  tibble_without_pathogen_spore_conc_column
)
}
\arguments{
\item{tibble_without_pathogen_spore_conc_column}{a tibble}
}
\value{
a tibble
}
\description{
add_pathogen_spore_conc_column_and_relocate
}
