% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{change_min_to_hour_in_reshape_image_classification_format}
\alias{change_min_to_hour_in_reshape_image_classification_format}
\title{change_min_to_hour_in_reshape_image_classification_format}
\usage{
change_min_to_hour_in_reshape_image_classification_format(
  tibble_with_min_reshape_format
)
}
\arguments{
\item{tibble_with_min_reshape_format}{a tibble}
}
\value{
a tibble
}
\description{
change_min_to_hour_in_reshape_image_classification_format
}
