% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{cleanup_time_rename_and_relocate_reshape_format}
\alias{cleanup_time_rename_and_relocate_reshape_format}
\title{cleanup_time_rename_and_relocate_reshape_format}
\usage{
cleanup_time_rename_and_relocate_reshape_format(
  tibble_with_long_reshape_format_uncleaned,
  sheet_string = "label"
)
}
\arguments{
\item{tibble_with_long_reshape_format_uncleaned}{a tibble}

\item{sheet_string}{a string}
}
\value{
a tibble
}
\description{
cleanup_time_rename_and_relocate_reshape_format
}
