% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abspath_col_derivations.R
\name{col_with_abspath_basename_sans_ext}
\alias{col_with_abspath_basename_sans_ext}
\title{col_with_abspath_basename_sans_ext}
\usage{
col_with_abspath_basename_sans_ext(tibble_with_abspath_col)
}
\arguments{
\item{tibble_with_abspath_col}{}
}
\description{
col_with_abspath_basename_sans_ext
}
