% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abspath_col_derivations.R
\name{col_with_abspath_first_directory_name}
\alias{col_with_abspath_first_directory_name}
\title{tibble_with_abspath_col}
\usage{
col_with_abspath_first_directory_name(tibble_with_abspath_col)
}
\arguments{
\item{tibble_with_abspath_col}{a tibble}
}
\value{
a tibble
}
\description{
tibble_with_abspath_col
}
