% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{convert_relocate_and_arrange_columns}
\alias{convert_relocate_and_arrange_columns}
\title{convert_relocate_and_arrange_columns}
\usage{
convert_relocate_and_arrange_columns(
  tibble_pivotted_long_reshape_preinc_joined
)
}
\arguments{
\item{tibble_pivotted_long_reshape_preinc_joined}{a tibble}
}
\value{
a tibble
}
\description{
convert_relocate_and_arrange_columns
}
