% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{convert_T_values_to_hour_and_relocate}
\alias{convert_T_values_to_hour_and_relocate}
\title{convert_T_values_to_hour_and_relocate}
\usage{
convert_T_values_to_hour_and_relocate(tibble_with_T_values_column)
}
\arguments{
\item{tibble_with_T_values_column}{a tibble}
}
\value{
a tibble
}
\description{
convert_T_values_to_hour_and_relocate
}
