% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{convert_well_to_row_col_and_relocate}
\alias{convert_well_to_row_col_and_relocate}
\title{convert_well_to_row_col_and_relocate}
\usage{
convert_well_to_row_col_and_relocate(tibble_with_well_column)
}
\arguments{
\item{tibble_with_well_column}{a tibble}
}
\value{
a tibble
}
\description{
convert_well_to_row_col_and_relocate
}
