% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondance_reshape_preincubation.R
\name{create_lookup_barcodes_per_preincubation_id}
\alias{create_lookup_barcodes_per_preincubation_id}
\title{create_lookup_barcodes_per_preincubation_id}
\usage{
create_lookup_barcodes_per_preincubation_id(events_preincubation)
}
\arguments{
\item{events_preincubation}{a tibble}
}
\value{
a tibble
}
\description{
create_lookup_barcodes_per_preincubation_id
}
