% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondance_reshape_preincubation.R
\name{create_lookup_barcodes_per_reshape_run}
\alias{create_lookup_barcodes_per_reshape_run}
\title{create_lookup_barcodes_per_reshape_run}
\usage{
create_lookup_barcodes_per_reshape_run(events_reshape)
}
\arguments{
\item{events_reshape}{a tibble}
}
\value{
a tibble
}
\description{
create_lookup_barcodes_per_reshape_run
}
