% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_partition.R
\name{create_partion_by_preincubation_id}
\alias{create_partion_by_preincubation_id}
\title{create_partion_by_preincubation_id}
\usage{
create_partion_by_preincubation_id(
  id_preinc,
  tibble_of_reshape_classifications_formatted_by_preinc_assay,
  path_folder_DTU_format_id_partion,
  extention_id_partion = ".xlsx"
)
}
\arguments{
\item{id_preinc}{a string}

\item{tibble_of_reshape_classifications_formatted_by_preinc_assay}{a tibble}

\item{path_folder_DTU_format_id_partion}{a string}

\item{extention_id_partion}{a string}
}
\value{
Nothing. Has file creation sideeffect
}
\description{
create_partion_by_preincubation_id
}
