% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondance_reshape_preincubation.R
\name{create_reshape_runs_with_preinc_id_and_plate_assay_role}
\alias{create_reshape_runs_with_preinc_id_and_plate_assay_role}
\title{create_reshape_runs_with_preinc_id_and_plate_assay_role}
\usage{
create_reshape_runs_with_preinc_id_and_plate_assay_role(
  events_reshape,
  events_preincubation
)
}
\arguments{
\item{events_reshape}{a tibble}

\item{events_preincubation}{a tibble}
}
\value{
a tibble
}
\description{
create_reshape_runs_with_preinc_id_and_plate_assay_role
}
