% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{find_xlsx_abspath_recursive_readxls_and_unnest}
\alias{find_xlsx_abspath_recursive_readxls_and_unnest}
\title{find_xlsx_abspath_recursive_readxls_and_unnest}
\usage{
find_xlsx_abspath_recursive_readxls_and_unnest(
  path_dir,
  sheet_string = "label"
)
}
\arguments{
\item{path_dir}{a string}

\item{sheet_string}{a string}
}
\value{
a tibble
}
\description{
find_xlsx_abspath_recursive_readxls_and_unnest
}
