% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{find_xlsxs_abspaths_recursive}
\alias{find_xlsxs_abspaths_recursive}
\title{find_xlsxs_abspaths_recursive}
\usage{
find_xlsxs_abspaths_recursive(path_dir)
}
\arguments{
\item{path_dir}{a string}
}
\value{
A named chr vector
}
\description{
find_xlsxs_abspaths_recursive
}
