% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_paths_to_tibble.R
\name{find_xlsxs_abspaths_recursive_and_put_in_column_in_tibble}
\alias{find_xlsxs_abspaths_recursive_and_put_in_column_in_tibble}
\title{find_xlsxs_abspaths_recursive_and_put_in_column_in_tibble}
\usage{
find_xlsxs_abspaths_recursive_and_put_in_column_in_tibble(path_dir)
}
\arguments{
\item{path_dir}{a string}
}
\value{
a tibble
}
\description{
find_xlsxs_abspaths_recursive_and_put_in_column_in_tibble
}
