% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{join_with_reshape_classification_and_remove_plate_level_details}
\alias{join_with_reshape_classification_and_remove_plate_level_details}
\title{join_with_reshape_classification_and_remove_plate_level_details}
\usage{
join_with_reshape_classification_and_remove_plate_level_details(
  tibble_reshape_runs_with_preinc_id_and_plate_assay_role,
  tibble_reshape_long_format
)
}
\arguments{
\item{tibble_reshape_runs_with_preinc_id_and_plate_assay_role}{a tibble}

\item{tibble_reshape_long_format}{a tibble}
}
\value{
a tibble
}
\description{
join_with_reshape_classification_and_remove_plate_level_details
}
