% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_partition.R
\name{partition_by_preincubation_id}
\alias{partition_by_preincubation_id}
\title{partition_by_preincubation_id}
\usage{
partition_by_preincubation_id(
  tibble_of_reshape_classifications_formatted_by_preinc_assay,
  path_folder_DTU_format_id_partion
)
}
\arguments{
\item{tibble_of_reshape_classifications_formatted_by_preinc_assay}{a tibble}

\item{path_folder_DTU_format_id_partion}{a string}
}
\value{
Nothing. Has file creation sideeffect
}
\description{
partition_by_preincubation_id
}
