% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{pivot_according_to_plate_assay_role}
\alias{pivot_according_to_plate_assay_role}
\title{pivot_according_to_plate_assay_role}
\usage{
pivot_according_to_plate_assay_role(
  tibble_with_reshape_classification_preinc_joined_format
)
}
\arguments{
\item{tibble_with_reshape_classification_preinc_joined_format}{a tibble}
}
\value{
a tibble
}
\description{
pivot_according_to_plate_assay_role
}
