% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{pivot_longer_reshape_image_classification_format}
\alias{pivot_longer_reshape_image_classification_format}
\title{pivot_longer_reshape_image_classification_format}
\usage{
pivot_longer_reshape_image_classification_format(
  tibble_with_wide_reshape_format
)
}
\arguments{
\item{tibble_with_wide_reshape_format}{a tibble}
}
\value{
a tibble
}
\description{
pivot_longer_reshape_image_classification_format
}
