% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{read_in_and_merge_reshape_labels_and_scores}
\alias{read_in_and_merge_reshape_labels_and_scores}
\title{read_in_and_merge_reshape_labels_and_scores}
\usage{
read_in_and_merge_reshape_labels_and_scores(path_reshape_scores)
}
\arguments{
\item{path_reshape_scores}{a string}
}
\value{
a tibble
}
\description{
read_in_and_merge_reshape_labels_and_scores
}
