% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{read_in_reshape_image_classification_format}
\alias{read_in_reshape_image_classification_format}
\title{read_in_reshape_image_classification_format}
\usage{
read_in_reshape_image_classification_format(
  path_reshape_scores,
  sheet_string = "label"
)
}
\arguments{
\item{path_reshape_scores}{a string}

\item{sheet_string}{a string}
}
\value{
a tibble
}
\description{
read_in_reshape_image_classification_format
}
