% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{reformat_according_to_plate_assay_role}
\alias{reformat_according_to_plate_assay_role}
\title{reformat_according_to_plate_assay_role}
\usage{
reformat_according_to_plate_assay_role(
  tibble_of_reshape_runs_with_preinc_id_and_plate_assay_role
)
}
\arguments{
\item{tibble_of_reshape_runs_with_preinc_id_and_plate_assay_role}{a tibble}
}
\value{
a tibble
}
\description{
reformat_according_to_plate_assay_role
}
