% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_find_and_read_to_tibble.R
\name{reformat_reshape_classifications_according_to_plate_assay_role}
\alias{reformat_reshape_classifications_according_to_plate_assay_role}
\title{reformat_reshape_classifications_according_to_plate_assay_role}
\usage{
reformat_reshape_classifications_according_to_plate_assay_role(
  reshape_classification,
  tibble_reshape_runs_with_preinc_id_and_plate_assay_role
)
}
\arguments{
\item{reshape_classification}{a tibble}

\item{tibble_reshape_runs_with_preinc_id_and_plate_assay_role}{a tibble}
}
\value{
a tibble
}
\description{
reformat_reshape_classifications_according_to_plate_assay_role
}
